Switch to enum over trait for layer
