+++ src/layers/linear.rs
@@ -16,18 +16,16 @@ impl Linear {
             bias: Array2::from_elem((input_shape, 1), 0.),
         }
     }
-}
 
-impl Layer for Linear {
-    fn forward(&self, input: Array2<f32>) -> Array2<f32> {
+    pub fn forward(&self, input: Array2<f32>) -> Array2<f32> {
         self.weights.dot(&input) + &self.bias
     }
 
-    fn backward(&self, activations: Array2<f32>) -> Array2<f32> {
+    pub fn backward(&self, activations: Array2<f32>) -> Array2<f32> {
         activations.dot(&self.weights.t())
     }
 
-    fn compute_bias_gradient(&self, activations: Array2<f32>) -> Array2<f32> {
+    pub fn compute_bias_gradient(&self, activations: Array2<f32>) -> Array2<f32> {
         activations.sum_axis(Axis(1)).insert_axis(Axis(1))
     }
 }
