+++ src/layers/relu.rs
@@ -1,15 +1,13 @@
 use ndarray::Array2;
 
-use crate::layer::Layer;
-
 pub struct ReLU;
 
-impl Layer for ReLU {
-    fn forward(&self, input: Array2<f32>) -> Array2<f32> {
+impl ReLU {
+    pub fn forward(&self, input: Array2<f32>) -> Array2<f32> {
         input.mapv(|x| x.max(0.))
     }
 
-    fn backward(&self, activations: Array2<f32>) -> Array2<f32> {
+    pub fn backward(&self, activations: Array2<f32>) -> Array2<f32> {
         activations.mapv(|x| if x > 0. { 1. } else { 0. })
     }
 }
