+++ src/lib.rs
@@ -10,16 +10,20 @@ use std::{
 
 use std::sync::mpsc::Sender;
 
-pub mod neural_network;
+pub mod neural_network_old;
+use layer::Layer;
 use ndarray::{array, Array2};
 
-use crate::neural_network::{NeuralNetwork, NEURAL_NETWORK_MANAGER};
+use crate::neural_network_old::{NeuralNetwork_Old, NEURAL_NETWORK_MANAGER};
 
 pub mod layer;
 pub mod layers;
 use layers::linear::Linear;
 use layers::relu::ReLU;
 use layers::sigmoid::Sigmoid;
+
+pub mod neural_network;
+use neural_network::NeuralNetwork;
 
 const TICK_SPEED: u32 = 1;
 
@@ -36,21 +40,25 @@ pub fn main() {
 
 
 
-    let mut network = create_network(
+    let mut network = create_network(vec![
         Linear::new(20, 20), 
         ReLU, 
         Linear::new(20, 10),
         Sigmoid,
         Linear::new(10, 3),
-    );
+    ]);
 }
 
-pub fn init(input_count: usize, output_count: usize) -> NeuralNetwork {
-    let neural_network = NeuralNetwork::new(1., 0.1, vec![input_count, 5, 3, output_count]);
+pub fn create_network(layers: Vec<Layer>) -> NeuralNetwork_Old {
+    let neural_network = NeuralNetwork_Old::new()
+}
+
+pub fn init(input_count: usize, output_count: usize) -> NeuralNetwork_Old {
+    let neural_network = NeuralNetwork_Old::new(1., 0.1, vec![input_count, 5, 3, output_count]);
     neural_network
 }
 
-pub fn run_ticks(neural_network: &mut NeuralNetwork, inputs: Vec<f32>) {
+pub fn run_ticks(neural_network: &mut NeuralNetwork_Old, inputs: Vec<f32>) {
     let time_start = Instant::now();
 
     for tick in 0..50000 {
